% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Anderson}
\alias{Anderson}
\title{Hypothetical data for Alexander Anderson}
\format{An object of class \code{matrix} with 4 rows and 2 columns.}
\source{
\url{http://rcompanion.org/handbook/H_04.html}
}
\usage{
Anderson
}
\description{
A matrix of counts for students passing or failing a pesticide 
training course across four counties. Hypothetical data.
}
\keyword{datasets}
