% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{AndersonBias}
\alias{AndersonBias}
\title{Hypothetical data for Alexander Anderson with sex bias}
\format{
An object of class \code{data.frame} with 16 rows and 4 columns.
}
\source{
\url{http://rcompanion.org/handbook/H_06.html}
}
\usage{
AndersonBias
}
\description{
A data frame of counts for students passing or failing a  
pesticicde training course across four counties, with sex of students. 
Hypothetical data.
}
\keyword{datasets}
