% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{AndersonRainBarrel}
\alias{AndersonRainBarrel}
\title{Hypothetical data for Alexander Anderson on rain barrel installation}
\format{An object of class \code{matrix} with 2 rows and 2 columns.}
\source{
\url{http://rcompanion.org/handbook/H_05.html}
}
\usage{
AndersonRainBarrel
}
\description{
A matrix of paired counts for students planning to install rain
             barrels before and after a class. Hypothetical data.
}
\keyword{datasets}
