% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{AndersonRainGarden}
\alias{AndersonRainGarden}
\title{Hypothetical data for Alexander Anderson on rain garden installation}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 3 columns.
}
\source{
\url{http://rcompanion.org/handbook/H_05.html}
}
\usage{
AndersonRainGarden
}
\description{
A matrix of paired counts for students planning to install rain
             gardens before and after a class. Hypothetical data.
}
\keyword{datasets}
