% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{BobBelcher}
\alias{BobBelcher}
\title{Hypothetical data for ratings of instructors in unreplicated CBD}
\format{
An object of class \code{data.frame} with 40 rows and 3 columns.
}
\source{
\url{http://rcompanion.org/handbook/F_10.html}
}
\usage{
BobBelcher
}
\description{
A data frame of Likert responses for five
             instructors for each of 8 respondents. 
             Arranged in unreplicated complete block design.
             Hypothetical data.
}
\keyword{datasets}
