% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Breakfast}
\alias{Breakfast}
\title{Hypothetical data for students' breakfast habits and travel to school}
\format{
An object of class \code{table} with 3 rows and 5 columns.
}
\source{
\url{http://rcompanion.org/handbook/H_09.html}
}
\usage{
Breakfast
}
\description{
A two-dimensional contingency table, in which Breakfast 
             is an ordered nominal variable,
             and Travel is a non-ordered nominal variable.
             Hypothetical data.
}
\keyword{datasets}
