% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{BrendonSmall}
\alias{BrendonSmall}
\title{Hypothetical data for Brendon Small and company}
\format{
An object of class \code{data.frame} with 45 rows and 6 columns.
}
\source{
\url{http://rcompanion.org/handbook/I_10.html}
}
\usage{
BrendonSmall
}
\description{
A data frame of the intake of calories and sodium for students 
             in five classes. Hypothetical data.
}
\keyword{datasets}
