% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{BullyHill}
\alias{BullyHill}
\title{Hypothetical data for proportion of students passing a certification}
\format{An object of class \code{data.frame} with 12 rows and 5 columns.}
\source{
\url{http://rcompanion.org/handbook/J_02.html}
}
\usage{
BullyHill
}
\description{
A data frame of counts of students passing and failing.
             Hypothetical data.
}
\keyword{datasets}
