% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Catbus}
\alias{Catbus}
\title{Hypothetical data for Catbus and company}
\format{An object of class \code{data.frame} with 26 rows and 5 columns.}
\source{
\url{http://rcompanion.org/handbook/C_03.html}
}
\usage{
Catbus
}
\description{
A data frame of the number of steps taken by students in three
             classes. Hypothetical data.
}
\keyword{datasets}
