% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cldList.r
\name{cldList}
\alias{cldList}
\title{Compact letter display for lists of comparisons}
\usage{
cldList(
  formula = NULL,
  data = NULL,
  comparison = NULL,
  p.value = NULL,
  threshold = 0.05,
  print.comp = FALSE,
  remove.space = TRUE,
  remove.equal = TRUE,
  remove.zero = TRUE,
  swap.colon = TRUE,
  swap.vs = FALSE,
  ...
)
}
\arguments{
\item{formula}{A formula indicating the variable holding p-values and
the variable holding the comparisons. 
e.g. P.adj ~ Comparison.}

\item{data}{The data frame to use.}

\item{comparison}{A vector of text describing comparisons, 
with each element in a form similar to
"Treat.A - Treat.B = 0".  Spaces and "=" and "0"
are removed by default}

\item{p.value}{A vector of p-values corresponding to the comparisons
in the \code{comparison} argument}

\item{threshold}{The alpha value.  That is, the p-value below which the
comparison will be considered significant}

\item{print.comp}{If \code{TRUE}, prints out a data frame of the
modified text of the comparisons.  Useful for debugging}

\item{remove.space}{If \code{TRUE}, removes spaces from the text of the
comparisons}

\item{remove.equal}{If \code{TRUE}, removes "=" from the text of the
comparisons}

\item{remove.zero}{If \code{TRUE}, removes "0" from the text of the
comparisons}

\item{swap.colon}{If \code{TRUE}, swaps ":" with "-" in the text of the
comparisons}

\item{swap.vs}{If \code{TRUE}, swaps "vs" with "-" in the text of the
comparisons}

\item{...}{Additional arguments passed to
\code{multcompLetters}}
}
\value{
A data frame of group names, group separation letters,
        and monospaced separtions letters
}
\description{
Produces a compact letter display (cld) from pairwise 
             comparisons that were summarized in a table of comparisons
}
\details{
The input should include either \code{formula} and \code{data};
          or \code{comparison} and \code{p.value}.
          
          This function relies upon the \code{multcompLetters}
          function in the \code{multcompView} package. The text for the
          comparisons
          passed to \code{multcompLetters} should be in the form
          "Treat.A-Treat.B".  Currently \code{cldList} removes
          spaces, equal signs, and zeros, by default, 
          and so can use 
          text in the form e.g.
          "Treat.A - Treat.B = 0".
          It also changes ":" to "-", and so can use
          text in the form e.g.
          "Treat.A : Treat.B".
}
\note{
The parsing of the formula is simplistic. 
         The first variable on the
         left side is used as the measurement variable.  
         The first variable on the
         right side is used for the grouping variable.
}
\examples{
data(PoohPiglet)
PoohPiglet$Speaker = factor(PoohPiglet$Speaker,
                            levels=c("Pooh", "Tigger", "Piglet"))
library(FSA)
DT = dunnTest(Likert ~ Speaker,
              data=PoohPiglet,
              method="bh")
DT = DT$res
DT
cldList(P.adj ~ Comparison,
        data      = DT,
        threshold = 0.05)

}
\references{
\url{http://rcompanion.org/handbook/F_08.html}
}
\author{
Salvatore Mangiafico, \email{mangiafico@njaes.rutgers.edu}
}
\concept{compact letter display cld post-hoc}
