% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{HayleySmith}
\alias{HayleySmith}
\title{Hypothetical data for responses about adopting lawn care practices}
\format{
An object of class \code{data.frame} with 56 rows and 3 columns.
}
\source{
\url{http://rcompanion.org/handbook/H_05.html}
}
\usage{
HayleySmith
}
\description{
A data frame in long form with yes/no responses for four
             lawn care practices for each of 14 respondents.
             Hypothetical data.
}
\keyword{datasets}
