% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Monarchs}
\alias{Monarchs}
\title{Hypothetical data for monarch butterflies in gardens}
\format{An object of class \code{data.frame} with 24 rows and 2 columns.}
\source{
\url{http://rcompanion.org/handbook/J_01.html}
}
\usage{
Monarchs
}
\description{
A data frame of the number of monarch butterflies in three
             gardens. Hypothetical data.
}
\keyword{datasets}
