% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Nurseries}
\alias{Nurseries}
\title{Data for proportion of good practices followed by plant nuseries}
\format{
An object of class \code{data.frame} with 38 rows and 2 columns.
}
\source{
Mangiafico, S.S., Newman, J.P., Mochizuki, M.J.,
        and Zurawski, D. (2008). Adoption of sustainable practices
        to protect and conserve water resources in container nurseries
        with greenhouse facilities. Acta horticulturae 797, 367-372.
}
\usage{
Nurseries
}
\description{
A data frame with two variables: size of plant nursery in
             hectares, and proportion of good practices followed by
             the nursery
}
\keyword{datasets}
