% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseMcnemar.r
\name{pairwiseMcnemar}
\alias{pairwiseMcnemar}
\title{Pairwise McNemar and related tests for Cochran Q test post-hoc}
\usage{
pairwiseMcnemar(x, g, block, test = "exact", method = "fdr", digits = 3,
  correct = FALSE)
}
\arguments{
\item{x}{The response variable.}

\item{g}{The grouping variable.}

\item{block}{The blocking variable.}

\item{test}{If \code{"exact"}, conducts an exact test of symmetry
analogous to a McNemar test.
If \code{"mcnemar"}, conducts a McNemar test of symmetry.
If \code{"permutation"}, conducts a permutation test 
analogous to a McNemar test.}

\item{method}{The method for adjusting multiple p-values.
See \code{\link{p.adjust}}.}

\item{digits}{The number of significant digits in the output.}

\item{correct}{If \code{TRUE}, applies a continuity correction
for the McNemar test.}
}
\value{
A list containing: a data frame of results of the global test;
        a data frame of results of the pairwise results;
        and a data frame mentioning the p-value adjustment method.
}
\description{
Conducts pairwise McNemar, exact, and permutation
             tests as a post-hoc to Cochran Q test.
}
\details{
The component tables for the pairwise tests
         must be of size 2 x 2.
}
\examples{
### Cochran Q post-hoc example
data(HayleySmith)
library(RVAideMemoire)
cochran.qtest(Response ~ Practice | Student,
              data = HayleySmith)
pairwiseMcnemar(x       = HayleySmith$Response,
                g       = HayleySmith$Practice,
                block   = HayleySmith$Student,
                test    = "exact",
                method  = "fdr",
                digits  = 3)
}
\author{
Salvatore Mangiafico, \email{mangiafico@njaes.rutgers.edu}
}
\references{
\url{http://rcompanion.org/handbook/H_07.html}
}
\seealso{
\code{\link{nominalSymmetryTest}}, \code{\link{groupwiseCMH}},
         \code{\link{pairwiseNominalIndependence}}, 
         \code{\link{pairwiseNominalMatrix}}
}
\concept{
McNemar Cochran Q permutation exact
}

