% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisePermutationSymmetryMatrix.r
\name{pairwisePermutationSymmetryMatrix}
\alias{pairwisePermutationSymmetryMatrix}
\title{Pairwise two-sample permutation symmetry 
       tests with matrix output}
\usage{
pairwisePermutationSymmetryMatrix(x, g, b, method = "fdr", ...)
}
\arguments{
\item{x}{The response variable as a vector.}

\item{g}{The grouping variable as a vector.}

\item{b}{The blocking variable as a vector.}

\item{method}{The p-value adjustment method to use for multiple tests.
See \code{\link{p.adjust}}.}

\item{...}{Additional arguments passed to
\code{\link{symmetry_test}}.}
}
\value{
A list consisting of:
        A matrix of p-values;
        the p-value adjustment method;
        a matrix of adjusted p-values.
}
\description{
Conducts pairwise two-sample permutation tests
             for symmetry across groups.
}
\details{
Permutation tests are non-parametric tests 
         that do not assume normally-distributed errors.
         See \url{http://rcompanion.org/rcompanion/d_06a.html} for
         futher discussion of this test.

         The \code{pairwisePermutationSymmetryMatrix} function
         can be used as a post-hoc method following an omnibus 
         permutation test analogous to a paired one-way analysis
         of variance.
         The matrix output can be converted to a compact letter display.
}
\examples{
data(BobBelcher)
BobBelcher = BobBelcher[order(factor(BobBelcher$Instructor, 
                        levels=c("Linda Belcher", "Louise Belcher",
                                 "Tina Belcher", "Bob Belcher",
                                 "Gene Belcher"))),]
BobBelcher$Likert.f = factor(BobBelcher$Likert, ordered=TRUE)
PT = pairwisePermutationSymmetryMatrix(x      = BobBelcher$Likert.f,
                                       g      = BobBelcher$Instructor,
                                       b      = BobBelcher$Rater,
                                       method = "fdr")$Adjusted
PT
library(multcompView)
multcompLetters(PT,
                compare="<",
                threshold=0.05,
                Letters=letters)   

}
\author{
Salvatore Mangiafico, \email{mangiafico@njaes.rutgers.edu}
}
\references{
\url{http://rcompanion.org/handbook/K_03.html}
}
\seealso{
\code{\link{pairwisePermutationSymmetry}}
}
\concept{
permutation nonparametric post-hoc one-way cld symmetry
}

