% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Pennsylvania18}
\alias{Pennsylvania18}
\title{Votes for the Democratic candidate in Pennsylvania 18 in 2016 and 2018}
\format{An object of class \code{matrix} with 2 rows and 2 columns.}
\source{
\url{http://rcompanion.org/handbook/H_10.html}
}
\usage{
Pennsylvania18
}
\description{
A two-by-two matrix with the proportion of votes for the
             Democratic candidate in two races, in 2016 and 2018.
             2016 is the Presidential election with Hilary Clinton as the
             Democratic candidate.  2018 is a House of Representatives
             election with Conor Lamb.  These data are for
             Pennsylvania's 18th Congressional District.
}
\keyword{datasets}
