% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Pooh}
\alias{Pooh}
\title{Hypothetical data for paired ratings of Pooh Bear}
\format{
An object of class \code{data.frame} with 20 rows and 4 columns.
}
\source{
\url{http://rcompanion.org/handbook/F_06.html}
}
\usage{
Pooh
}
\description{
A data frame of Likert responses for instructor Pooh Bear
             for each of 10 respondents, paired before and after. 
             Hypothetical data.
}
\keyword{datasets}
