% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{PoohPiglet}
\alias{PoohPiglet}
\title{Hypothetical data for ratings of Pooh, Piglet, and Tigger}
\format{
An object of class \code{data.frame} with 30 rows and 2 columns.
}
\source{
\url{http://rcompanion.org/handbook/F_08.html}
}
\usage{
PoohPiglet
}
\description{
A data frame of Likert responses for instructors Pooh Bear,
             Piglet, and Tigger.
             Hypothetical data.
}
\keyword{datasets}
