% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Religion}
\alias{Religion}
\title{Hypothetical data for change in religion after a caucusing event}
\format{An object of class \code{matrix} with 4 rows and 4 columns.}
\source{
\url{http://rcompanion.org/handbook/H_05.html}
}
\usage{
Religion
}
\description{
A matrix of paired counts for religion of people before 
             and after an event.
             Hypothetical data.
}
\keyword{datasets}
