% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxonPS.r
\name{wilcoxonPS}
\alias{wilcoxonPS}
\title{Grissom and Kim's Probability of Superiority (PS)}
\usage{
wilcoxonPS(
  formula = NULL,
  data = NULL,
  x = NULL,
  y = NULL,
  ci = FALSE,
  conf = 0.95,
  type = "perc",
  R = 1000,
  histogram = FALSE,
  digits = 3,
  reportIncomplete = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{formula}{A formula indicating the response variable and
the independent variable. e.g. y ~ group.}

\item{data}{The data frame to use.}

\item{x}{If no formula is given, the response variable for one group.}

\item{y}{The response variable for the other group.}

\item{ci}{If \code{TRUE}, returns confidence intervals by bootstrap.
May be slow.}

\item{conf}{The level for the confidence interval.}

\item{type}{The type of confidence interval to use.
Can be any of "\code{norm}", "\code{basic}", 
              "\code{perc}", or "\code{bca}".
Passed to \code{boot.ci}.}

\item{R}{The number of replications to use for bootstrap.}

\item{histogram}{If \code{TRUE}, produces a histogram of bootstrapped values.}

\item{digits}{The number of significant digits in the output.}

\item{reportIncomplete}{If \code{FALSE} (the default),
\code{NA} will be reported in cases where there
are instances of the calculation of the statistic
failing during the bootstrap procedure.}

\item{verbose}{If \code{TRUE}, reports the proportion of ties and
the proportions of (Ya > Yb) and (Ya < Yb).}

\item{...}{Additional arguments, not used.}
}
\value{
A single statistic, PS.
        Or a small data frame consisting of PS,
        and the lower and upper confidence limits.
}
\description{
Calculates Grissom and Kim's Probability of Superiority (PS)
             with confidence intervals by bootstrap
}
\details{
PS is an effect size statistic appropriate
         in cases where a Wilcoxon-Mann-Whitney test might be used.
         It ranges from 0 to 1, with 0.5 indicating stochastic equality,
         and 1 indicating that the first group dominates the second.
         
         PS is defined as P(Ya > Yb), with no provision made for tied values
         across groups.
         
         If there are no tied values, PS will be equal to VDA. 
           
         The input should include either \code{formula} and \code{data};
         or \code{x}, and \code{y}. If there are more than two groups,
         only the first two groups are used.
         
          Currently, the function makes no provisions for \code{NA}
          values in the data.  It is recommended that \code{NA}s be removed
          beforehand.
          
          When the data in the first group are greater than
          in the second group, PS is greater than 0.5.
          When the data in the second group are greater than
          in the first group, PS is less than 0.5.
          
          Be cautious with this interpretation, as R will alphabetize
          groups in the formula interface if the grouping variable
          is not already a factor.

          When PS is close to 0 or close to 1,
          or with small sample size,
          the confidence intervals 
          determined by this
          method may not be reliable, or the procedure may fail.
}
\note{
The parsing of the formula is simplistic. 
         The first variable on the
         left side is used as the measurement variable.  
         The first variable on the
         right side is used for the grouping variable.
}
\examples{
data(Catbus)
wilcoxonPS(Steps ~ Sex, data=Catbus, verbose=TRUE)

}
\references{
Grissom, R.J. and J.J. Kim. 2012. Effect Sizes for Research. 2nd ed. 
            Routledge, New York.
            
            \url{http://rcompanion.org/handbook/F_04.html}
}
\seealso{
\code{\link{cliffDelta}}, \code{\link{vda}}
}
\author{
Salvatore Mangiafico, \email{mangiafico@njaes.rutgers.edu}
}
\concept{effect size}
