% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nagelkerkeHermite.r
\name{nagelkerkeHermite}
\alias{nagelkerkeHermite}
\title{[Defunct!] Pseudo r-squared measures for hermite models}
\usage{
nagelkerkeHermite(...)
}
\arguments{
\item{...}{Anything.}
}
\description{
Defunct. Produces McFadden, Cox and Snell, and Nagelkerke pseudo 
             R-squared measures, along with p-value for the model, 
             for hermite regression objects.
}
