% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseDifferences.r
\name{pairwiseDifferences}
\alias{pairwiseDifferences}
\title{[Defunct!] Pairwise differences for unreplicated CBD}
\usage{
pairwiseDifferences(...)
}
\arguments{
\item{...}{Anything.}
}
\description{
Defunct. Calculates the differences in the response variable
             for each pair of levels of a grouping variable 
             in an unreplicated complete block design.
}
