% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseOrdinalMatrix.r
\name{pairwiseOrdinalMatrix}
\alias{pairwiseOrdinalMatrix}
\title{[Defunct!] Pairwise two-sample ordinal regression with matrix output}
\usage{
pairwiseOrdinalMatrix(...)
}
\arguments{
\item{...}{Anything.}
}
\description{
Defunct. Performs pairwise two-sample ordinal regression 
             across groups.
}
