% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseOrdinalPairedMatrix.r
\name{pairwiseOrdinalPairedMatrix}
\alias{pairwiseOrdinalPairedMatrix}
\title{[Defunct!] Pairwise two-sample ordinal regression for paired data
       with matrix output}
\usage{
pairwiseOrdinalPairedMatrix(...)
}
\arguments{
\item{...}{Anything.}
}
\description{
Defunct. Performs pairwise two-sample ordinal regression 
             across groups for paired data
             with matrix output.
}
