% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseOrdinalPairedTest.r
\name{pairwiseOrdinalPairedTest}
\alias{pairwiseOrdinalPairedTest}
\title{[Defunct!] Pairwise two-sample ordinal regression for paired data}
\usage{
pairwiseOrdinalPairedTest(...)
}
\arguments{
\item{...}{Anything.}
}
\description{
Defunct. Performs pairwise two-sample ordinal regression 
             across groups for paired data.
}
