% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseOrdinalTest.r
\name{pairwiseOrdinalTest}
\alias{pairwiseOrdinalTest}
\title{[Defunct!] Pairwise two-sample ordinal regression}
\usage{
pairwiseOrdinalTest(...)
}
\arguments{
\item{...}{Anything.}
}
\description{
Defunct. Performs pairwise two-sample ordinal regression 
             across groups.
}
