% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseRobustMatrix.r
\name{pairwiseRobustMatrix}
\alias{pairwiseRobustMatrix}
\title{[Defunct!] Pairwise two-sample robust tests with matrix output}
\usage{
pairwiseRobustMatrix(...)
}
\arguments{
\item{...}{Anything.}
}
\description{
Defunct. Performs pairwise two-sample robust tests across groups
             with matrix output.
}
