% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseRobustTest.r
\name{pairwiseRobustTest}
\alias{pairwiseRobustTest}
\title{[Defunct!] Pairwise two-sample robust tests}
\usage{
pairwiseRobustTest(...)
}
\arguments{
\item{...}{Anything.}
}
\description{
Defunct. Performs pairwise two-sample robust tests across groups.
}
