% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseSignMatrix.r
\name{pairwiseSignMatrix}
\alias{pairwiseSignMatrix}
\title{[Defunct!] Pairwise sign tests with matrix output}
\usage{
pairwiseSignMatrix(...)
}
\arguments{
\item{...}{Anything.}
}
\description{
Defunct. Performs pairwise sign tests.
}
