% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseSignTest.r
\name{pairwiseSignTest}
\alias{pairwiseSignTest}
\title{[Defunct!] Pairwise sign tests}
\usage{
pairwiseSignTest(...)
}
\arguments{
\item{...}{Anything.}
}
\description{
Defunct. Performs pairwise sign tests.
}
